% Generated by roxygen2 (4.0.1): do not edit by hand
\name{github_update}
\alias{github_update}
\title{github_update - find, report and optionally update packages installed from or available on github}
\usage{
github_update(auto.install = FALSE, .progress = FALSE)
}
\arguments{
\item{auto.install}{should it try to auto-install newer packages [not recommended] (bool, initially \code{FALSE})}

\item{.progress}{show a progress bar (bool, initially \code{FALSE})}
}
\value{
data frame of installed packages with information on owner & whether an update is available
}
\description{
Finds local packages that [may] have github versions, then informs &
    optionally updates them to the newest versions. It is recommended that you
    still perform the updates by hand since you may want to double-check
    the repo for any gotchas or incompatibilities with other installed packages.
}
\note{
package installation exceptions are caught with \code{try}, but are not reported out in any other way but console errors
}
\examples{
\dontrun{
github_update()
##      package.repo       owner installed.version current.version update.available
## 1        corrplot      taiyun              0.73            0.73               no
## 2      data.table  Rdatatable             1.9.3           1.9.3               no
## 3          gmailr   jimhester             0.0.1           0.0.1               no
## 4        miniCRAN      andrie            0.0-20          0.0-20               no
## 5   RcppArmadillo    RcppCore         0.4.400.0       0.4.400.0               no
## 6        corrplot      taiyun              0.73            0.73               no
## 7        forecast robjhyndman               5.4             5.6              yes
## 8         formatR       yihui              0.10          0.10.5              yes
## 9         ggplot2      hadley             1.0.0        1.0.0.99              yes
## 10       ggthemes      jrnold             1.7.0           1.8.0              yes
## 11          highr       yihui               0.3           0.3.1              yes
## 13       jsonlite  jeroenooms             0.9.9          0.9.10              yes
## 14          knitr       yihui             1.6.6          1.6.14              yes
## 15 knitrBootstrap   jimhester             1.0.0           1.0.0               no
## 16      lubridate      hadley             1.3.3           1.3.3               no
## 17       markdown     rstudio               0.7           0.7.4              yes
## 18        memoise      hadley             0.2.1          0.2.99              yes
## 19           mime       yihui             0.1.1           0.1.2              yes
## 20        packrat     rstudio             0.4.0        0.4.0.12              yes
## 21           Rcpp    RcppCore            0.11.2        0.11.2.1              yes
## 22       reshape2      hadley               1.4        1.4.0.99              yes
## 23           rzmq    armstrtw             0.7.0           0.7.0               no
## 24         scales      hadley             0.2.4        0.2.4.99              yes
## 25          shiny     rstudio       0.10.0.9001     0.10.1.9004              yes
## 26       shinyAce trestletech             0.1.0           0.1.0               no
## 27        slidify    ramnathv             0.4.5           0.4.5               no
## 28         testit       yihui               0.3             0.3               no
}
}

