% Generated by roxygen2 (4.0.1): do not edit by hand
\name{whats_new}
\alias{whats_new}
\title{whats_new - produce a report on the new features of any out of date packages}
\usage{
whats_new()
}
\description{
Uses \link{render} from \link{rmarkdown} to generate a report of \code{NEWS} files from any packages that have updates not currently installed
}

