% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_update.r
\name{github_update}
\alias{github_update}
\title{Find, report and optionally update packages installed from or available on github}
\usage{
github_update(auto.install = FALSE, ask = TRUE, widget = FALSE,
  dependencies = TRUE, libpath = .libPaths()[1])
}
\arguments{
\item{auto.install}{Should it try to auto-install newer packages [not
recommended] (bool, initially \code{FALSE})}

\item{ask}{If this and \code{auto.install}==\code{TRUE} a pkg install select
list will be shown (graphical or text widget dependent upon \code{widget})}

\item{widget}{If this and \code{auto.install}==\code{TRUE} and
\code{ask}==\code{TRUE} then a GUI select widget will}
}
\value{
\code{data.frame} of installed packages with information on owner &
  whether an update is available (optionally includes installed library path)
}
\description{
Finds local packages that [may] have github versions, then informs &
optionally updates them to the newest versions. It is recommended that you
still selectively perform the updates since you may want to double-check the
repo for any gotchas or incompatibilities with other installed packages.
}
\details{
Added code from Thomas J Leeper's (@thosjleeper) \code{#spiffy} gist:
\url{https://gist.github.com/leeper/9123584}
}
\note{
Package installation exceptions are caught with \code{try}, but are not
  reported out in any other way but console errors
}
