% Generated by roxygen2 (4.0.1): do not edit by hand
\name{github_update}
\alias{github_update}
\title{github_update - find, report and optionally update packages installed from or available on github}
\usage{
github_update(auto.install = FALSE, ask = TRUE, widget = FALSE,
  .progress = FALSE, all = TRUE, show.location = FALSE)
}
\arguments{
\item{auto.install}{should it try to auto-install newer packages [not recommended] (bool, initially \code{FALSE})}

\item{ask}{if this and \code{auto.install}==\code{TRUE} a pkg install select list will be shown (graphical or text widget dependent upon \code{widget})}

\item{widget}{if this and \code{auto.install}==\code{TRUE} and \code{ask}==\code{TRUE} then a GUI select widget will}

\item{.progress}{show a progress bar (bool, initially \code{FALSE})}

\item{all}{search all \code{.libPaths()} (\code{TRUE}) or only the first one (initially \code{TRUE})}

\item{show.location}{include the library location in the returned data frame (initially \code{FALSE})}
}
\value{
data frame of installed packages with information on owner & whether an update is available (optionally includes installed library path)
}
\description{
Finds local packages that [may] have github versions, then informs &
    optionally updates them to the newest versions. It is recommended that you
    still selectively perform the updates since you may want to double-check
    the repo for any gotchas or incompatibilities with other installed packages.

    Added code from Thomas J Leeper's (@thosjleeper) \code{#spiffy} gist: \url{https://gist.github.com/leeper/9123584}
}
\note{
package installation exceptions are caught with \code{try}, but are not reported out in any other way but console errors
}
\examples{
\dontrun{
github_update()

##           package           repo       owner installed.version current.version update.available
## 1            Rcpp           Rcpp    RcppCore            0.11.2        0.11.2.1             TRUE
## 2   RcppArmadillo  RcppArmadillo    RcppCore         0.4.400.0       0.4.400.0            FALSE
## 3      data.table     data.table  Rdatatable             1.9.3           1.9.3            FALSE
## 4        dtupdate       dtupdate    hrbrmstr               1.1             1.1            FALSE
## 5        forecast       forecast robjhyndman               5.4             5.6             TRUE
## 6         formatR        formatR       yihui              0.10          0.10.5             TRUE
## 7          gmailr         gmailr   jimhester             0.0.1           0.0.1            FALSE
## 8        jsonlite       jsonlite  jeroenooms             0.9.9          0.9.10             TRUE
## 9           knitr          knitr       yihui             1.6.6          1.6.14             TRUE
## 10 knitrBootstrap knitrBootstrap   jimhester             1.0.0           1.0.0            FALSE
## 11      lubridate      lubridate      hadley             1.3.3           1.3.3            FALSE
## 12       markdown       markdown     rstudio             0.7.4           0.7.4            FALSE
## 13        memoise        memoise      hadley             0.2.1          0.2.99             TRUE
## 14           mime           mime       yihui             0.1.2           0.1.2            FALSE
## 15       miniCRAN       miniCRAN      andrie            0.0-20          0.0-20            FALSE
## 16        packrat        packrat     rstudio             0.4.0        0.4.0.12             TRUE
## 17       reshape2        reshape      hadley               1.4        1.4.0.99             TRUE
## 18         resolv         resolv    hrbrmstr             0.2.2           0.2.2            FALSE
## 19          rzmq           rzmq    armstrtw             0.7.0           0.7.0            FALSE
## 20         scales         scales      hadley             0.2.4        0.2.4.99             TRUE
## 21          shiny          shiny     rstudio       0.10.0.9001     0.10.1.9004             TRUE
## 22       shinyAce       shinyAce trestletech             0.1.0           0.1.0            FALSE
## 23        slidify        slidify    ramnathv             0.4.5           0.4.5            FALSE
## 24         testit         testit       yihui               0.3             0.3            FALSE
}
}

